Generalized submodule code to work for all submodules
All submodules can now be commits in the user's local fork of the submodule
on GitLab.  The build will fetch the user's fork to get the correct code,
but a step at the end of the build will fail since the submodule has not been
merged into its origin/master branch.