Addresses issue #17.
GitlabMergeRequestPoller now puts back a link to CDash to refer to the
builds for that particular merge request.

Since pollers don't have access to properties, had to refactor the
mechanism to provide CDash url and project names. The cdash url is now
specified in secrets.json while the project names are specified in
poll.py modules for each of the projects. pollers.make_pollers() ensures
that the cdash project names and cdash url is passed to the pollers.