Refactoring factory logic.
Added mechanism to customize factories for each project. This became
essential since we need to add a custom factor for superbuild to ensure
it has two source build steps to checkout superbuild repo and checkout
paraview repo.

Also cleaned up issues with codebases argument passed to schedulers.
Since that argument is now necessary, it seems it also needs to be
initialized fully. Hence added a method projects.get_codebase() that
makes this easier.